+++ @es-labs/node/comms/telegram.js
@@ -1,33 +1,37 @@
 'use strict'
 
 const axios = require('axios')
+let apiKey
+let channelId
 
-const { TELEGRAM_API_KEY, TELEGRAM_CHANNEL_ID } = global.CONFIG
+function setup(options = global.CONFIG) {
+  const { TELEGRAM_API_KEY, TELEGRAM_CHANNEL_ID } = options || {}
+  // console.log('TELEGRAM_API_KEY, TELEGRAM_CHANNEL_ID', TELEGRAM_API_KEY, TELEGRAM_CHANNEL_ID)
+  apiKey = TELEGRAM_API_KEY
+  channelId = TELEGRAM_CHANNEL_ID
+}
 
-// console.log('TELEGRAM_API_KEY, TELEGRAM_CHANNEL_ID', TELEGRAM_API_KEY, TELEGRAM_CHANNEL_ID)
-
-async function sendTgChannel(text) {
+async function sendChannelMsg(text) {
   try {
-    const rv = await axios.get('https://api.telegram.org/bot' + TELEGRAM_API_KEY + '/sendMessage?chat_id=' + TELEGRAM_CHANNEL_ID + '&text=' + text)
-    console.log()
-    return rv // { id, date. pts, seq }
+    const rv = await axios.get('https://api.telegram.org/bot' + apiKey + '/sendMessage?chat_id=' + channelId + '&text=' + text)
+    return rv // { id, date, pts, seq }
   } catch (e) {
     return { err: e.toString() }
   }
 }
 
-async function sendTgChatId(chatId, text) {
+async function sendChatMsg(chatId, text) {
   try {
     // console.log('chatId, text', chatId, text)
-    const rv = await axios.get('https://api.telegram.org/bot' + TELEGRAM_API_KEY + '/sendMessage?chat_id=' + chatId + '&text=' + text)
-    console.log()
-    return rv // { id, date. pts, seq }
+    const rv = await axios.get('https://api.telegram.org/bot' + apiKey + '/sendMessage?chat_id=' + chatId + '&text=' + text)
+    return rv // { id, date, pts, seq }
   } catch (e) {
     return { err: e.toString() }
   }
 }
 
 module.exports = {
-  sendTgChannel,
-  sendTgChatId
+  setup,
+  sendChannelMsg,
+  sendChatMsg
 }